rename + recording + xml + misc
