+++ .gitignore
@@ -1,0 +1,2 @@
+*.tgz
+mission_records/*
